chore: refactor create cookie header value
